POC - simplify: consolidated service and persistence modules
