+++ service/src/main/java/gov/va/vro/persistence/model/VeteranEntity.java
