+++ settings.gradle
@@ -76,7 +76,6 @@ include ':container-init'
 include ':docs'
 include ':db-init'
 include ':opa-init'
-include ':persistence:model'
 include ':scripts'
 include ':service'
 include ':service-python:assessclaimdc7101'
